zeroCrossings
This function calculates the zeros of a discrete function and some of its properties. 
Given a list of {{x_1,y_1}, {x_2, y_2}, ..., {x_N,y_N}}, it returns a list of properties of the M fixed points found in the form {{xs_1,dx_1,i_1},...,{xs_M,dx_M,i_M}}, where xs_k, is an estimation of the root k, dx_k, the derivative at that point, and i_k the index corresponding to the closest x_j.